Fix build for .NET 3.5
Code was recently added to allow pulling a new AggregateDictionary from the container.
It doesn't work in .NET 3.5.
It also *should* be unnecessary. The problem was raised because Spark views
are not getting access to the container of the currently executing request,
and therefore do not already have an AggregateDictionary.