Modified FieldAccessService to construct element requests with a service locator.
ElementRequest has a method Get<T> that uses it's service locator dependency.  You should be able to call that method without getting a null reference excpetion.  When field access service was constructing element requests, it was passing null into the service locator dependency.  FieldAccessService took on a dependency for service locator and is now passing that into the constructor for element requests.

I don't think it should be valid to construct an element request with a null service locator.  Consumers should be able to assume they can call Get<T> on element request always.  Will talk with the guys about it.