Added a more robust numeric type family
ie; If you register a string conversion for int in your fubu registry

> this.StringConversions(x =>
> 	{
>		x.IfIsType<int>().ConvertBy(i => i.ToString("N"));
>	});

If you were to have a value >= 1000 and run that through your HtmlConventions

The conversion lambda will likely produce an input tag with the value

"1,000"

"1,000" is a *valid* string representation of an integer but the .NET frameworks built in TypeConverter will not accept a string with a comma as a valid number. It will throw an Invalid Format exception.

So the new NumericTypeFamily will check the current thread culture and strip off the NumberGroupSeparator just before converting the value.