Register packages with the virtual path provider in alphabetical order
The virtual path provider allows us to overrides views in the application
with views in a package. However, we also want to override a package with
another package, so the order the packages are checked is important.

The cheap answer is to sort the packages alphabetically by name.
A longer term answer might involve adding more metadata to the package
manifest to describe its precedence.