Add how to contribute sections
The focus of this is to give people considering contributing to FubuMVC
some guidance on what they can do to help the project. Most of the
documentation here is on actually improving the documentation itself.
Eventually we'll also add committing code to FubuMVC but for now that is
handled by the fubu-project.org Contributing section [1]

* Included cross-references to Sphinx documentation

* Updated the index.rst to include documents in the internals/
  directory.

* Added a section on contributing to FubuMVC

* Added a section on contributing documentation to FubuMVC

* Added a section on where to get started with the documentation effort.

* Provided some guidance on how to use Sphinx and reStructuredText for
  documenting the code. (Borrowing generously from Django's
  documentation, thanks Django team!)

On a sidenote, eventually this internals documentation will eventually
be a superset of what is on the contribute page of fubu-project.org. In
here we'll include all the nitty gritty details of contributing code
such as naming conventions, whitespace preference, commit messages, etc.

[1] http://fubu-project.org/contribute