Update contributing-documentation
Fixed a typo in the 'do you have any guides' heading.

Also added a quick bullet point on looking through the test code to get
ideas on what to document. This section might need to get expanded a
little more but it'll help everyone new to FubuMVC to see what's
available.