Fixed image include in usingspark.rst
Missed a backslash in the .. image include in the usingspark.rst.
Apparently it will work just fine on Windows when you build it locally
but when it ends up on ReadTheDocs the URL to the image is broken
because of the backslash. So make sure you always use forward slashes
when doing file paths in reStructuredText. Literal includes, images,
etc.