Added more to the writing documentation guide
Added the meaning of the folder structure under the docs/ folder as well
as trying to flesh out more of the conventions for writing the
documentation itself. Since Sphinx and reStructuredText treat whitespace
as significant I've suggested we use only spaces for rst files. The more
consistent we can get the format and layout of our docs among the
contributors the better.