Internet Explorer does not support type="application/javascript" for script tags
Clearly IE is so pants on head retarded that we don't even bother
to open it and test things anymore.

I spent a couple of hours punching myself in the dick trying to use
IE's developer tools trying to figure out why $ was undefined

After putting alert("say something motha !@#$a!"); in jquery.js
and not getting anything back and raging (╯°□°）╯︵ ┻━┻
flipping tables I saw the application/javascript and thought
maybe its that and it was.

Silly IE when will you finally die