Adding deployment topic
Forgot to add the deployment topic because it was being gitignored.
Updating the gitignore to negate that directory since I don't know what
implications there would be to changing the other 5 lines that have some
variation of deployment in it.

Moved all the gitignores involving deployment next to each other.